searchengine-webapi/requirements.txt
requests~=2.26.0
Flask~=2.0.0
flask-restful~=0.3.0
pyyaml~=5.4.0
elasticsearch~=7.14.0
pyjwt[crypto]~=2.1.0